Fix for issue 48 ("JSONValue.parse does not correctly decode UTF-8 bytes from an inputstream", https://code.google.com/p/json-smart/issues/detail?id=48).
Instead of taking a byte at a time from the InputStream, this now wraps
the InputStream in an InputStreamReader and uses the existing
JSONParserReader to read one character at a time.
    
A few extra parse() methods were added to take the source bytes'
character set as a parameter. This allows for cases when the platform's
default character set is not wanted.